+++ NEWS.md
@@ -1,5 +1,14 @@
 # rsconnect 0.8.30 (development version)
 
+* `deployApp()` now derives `appName` from `appDir` and `appPrimaryDoc`, 
+  never using the title. It also tells you what name has used (#538).
+  It now only simplifies the path if you are publishing to shinyapps.io,
+  since its restrictions on application names are much tighter than those
+  of Posit Connect.
+  
+* `generateAppName()` has been deprecated as using the `appTitle` to generate
+  the `appName` leads to confusion if you later try to change the title.
+
 * `deployApp()` is more aggressive about saving deployment data, which should
   make it less likely that you need to repeat yourself after a failed 
   deployment. In particular, it now saves both before and after uploading the